refactor slashing code and fix audit todos
